Fix server disappearing from Steam CS:GO browser tab
Root cause: AppID was stored as `short` (16-bit signed, max 32767) and
hardcoded to 730. Post-CS2 CS:GO servers use AppID 4465480 which
overflows `short`, causing Steam to not recognize the game.

Fixes:
- Change m_iAppID from `short` to `int` (supports any AppID)
- Change SetAppID/GetAppID signatures from short to int
- Read AppID from steam.inf instead of hardcoding 730
- Handle AppID > 65535 in body field per A2S_INFO protocol spec
  (write 0 in 16-bit body field, full value in 64-bit GameID EDF)
- Cast to uint64_t for WriteLongLong GameID EDF field

+++ natives.h
@@ -110,8 +110,8 @@ public:
     void SetGameDiscription(const char* sName, bool bDefault = false);
     const char* GetGameDiscription();
     
-    void SetAppID(short iId, bool bDefault = false);
-    short GetAppID();
+    void SetAppID(int iId, bool bDefault = false);
+    int GetAppID();
     
     void SetNumClients(uint8_t iClientCount, bool bDefault = false);
     uint8_t GetNumClients();
@@ -168,7 +168,7 @@ private:
     char m_GameDescription[256];
     bool m_bDefaultGameDescription;
     
-    short m_iAppID;
+    int m_iAppID;
     bool m_bDefaultAppID;
     
     uint8_t m_iNumClients;
